Explicit Python paths (basic users likely won't have python in path)
Since the installer in the Dependencies folder isn't configured to install this program's Python instance into PATH, we can't expect it to be there. This repo should be optimized for more basic and out-of-the-box operation. On the other hand, more advanced users should be able to easily find and change these hardcoded filepaths if needed.