More intelligent default part selection in Model tab
Refers to the Model tab of the Texture Editor interface. Now attempts to intelligently determine a model part (DObj) to show by default in the Model tab when a texture is selected. Prioritizes selection for high or low-poly parts depending on the current state of that option. If a part is not found for the current high/low-poly option, the option will be changed.